refactor: move shared general tab functionality to base classes.
